+++ src/main/java/io/elice/shoppingmall/product/Entity/Option/Gender.java
@@ -1,6 +1,6 @@
 package io.elice.shoppingmall.product.Entity.Option;
 
-import io.elice.shoppingmall.entity.base.BaseEntity;
+import io.elice.shoppingmall.entity.baseEntity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -13,7 +13,7 @@ import lombok.*;
 public class Gender extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column
     private String genderName;
